+++ src/guifuncs.h
@@ -24,5 +24,6 @@ struct Image *MenuImage(CONST_STRPTR name, struct Screen *);
 void windowBlocking(struct Window *, Object *, BOOL);
 void gadgetBlocking(struct Window *, Object *, BOOL);
 void showMsgReq(Object *, CONST_STRPTR, CONST_STRPTR);
+int listCount(struct List *);
 
 #endif